% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clickableImageMap.R
\name{calibrate}
\alias{calibrate}
\title{calibrate}
\usage{
calibrate(m, rows, pullDownRow)
}
\arguments{
\item{m}{matrix}

\item{rows}{list of row heights in the gtable object}

\item{pullDownRow}{integer number of the target row in the gtable object}
}
\value{
returns a list whose components are matrices
representing the upper and lower coordinates of the
row and column cells
}
\description{
use coordinates of upper left and bottom right of matrix
to construct mapping between viewport coordinates and matrix cells
}
\examples{
if(interactive()){
m<-matrix(1:20 * .05,nrow=2,ncol=10)
load("data/x_rows.RData")
pullDownRow<-which(names(x_rows)=="pullDownRow")
load("data/x_m.RData")
cal<-calibrate(x_m,x_rows,pullDownRow)
}

}
