% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clickableImageMap.R
\name{clickableImageMapDemo}
\alias{clickableImageMapDemo}
\title{clickableImageMapDemo}
\usage{
clickableImageMapDemo(
  n = 3,
  bounds = list(xmin = 0.534, xmax = 0.553, ymin = 0.057, ymax = 0.067),
  sleepTime = 0.5
)
}
\arguments{
\item{n}{integer number of values to be edited in matrix m}

\item{bounds}{list of 4 numerical values xmin, xmax, ymin, ymax}

\item{sleepTime}{numeric number of seconds to sleep to avoid potential race condition}
}
\value{
returns the updated numerical matrix
}
\description{
demo to illustrate how to implement calibrate()
and grid.locator() for a numerical matrix. This is just a stub
to be replaced by the user's actual program.
}
\details{
this package emulates edit() but allows full control
over formatting and management of the edited matrix.
sleepTime parameter can be set to nonzero (suggest trying sleepTime=0.5) in case a complicated \cr
graphic causes a race condition evidenced by incomplete redrawing of the window. Too large a value \cr
might cause a noticeable annoying delay in redrawing the window.
}
\examples{
if(interactive()){
m<-clickableImageMapDemo(2,bounds=list(xmin=.534,xmax=.553,ymin=.057,ymax=.067))
}

}
