% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clickableImageMap.R
\name{construct_entire_gtab}
\alias{construct_entire_gtab}
\title{construct_entire_gtab}
\usage{
construct_entire_gtab(m, rows, message, clickCoord)
}
\arguments{
\item{m}{a matrix}

\item{rows}{numerical vector defining rows for inserting grobs into main gtable}

\item{message}{character string message to display in annunciator grob of gtable}

\item{clickCoord}{numerical matrix of 2 columns, each row contains x and y coords
of a mouse click}
}
\value{
returns a list whose components are
\itemize{
\item m.pullDown component m of return value of pullDown()
\item cal.pullDown return value of calibrate()
\item cal.m return value of calibrate()
\item gtab return value of annunciator()
}
}
\description{
construct the main gtable into which grobs will be inserted
}
\examples{
if(interactive()){
load("data/x_m.RData")
load("data/x_rows.RData")
load("data/x_clickCoord.RData")
gtab<-construct_entire_gtab(x_m,x_rows,"x_message",x_clickCoord)
}

}
