% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clickableImageMap.R
\name{decode}
\alias{decode}
\title{decode}
\usage{
decode(y, cal, rcnames)
}
\arguments{
\item{y}{parsed return value of grid.locator()}

\item{cal}{return value of calibrate()}

\item{rcnames}{Boolean if TRUE matrix has row names and col names}
}
\value{
returns an integer vector of the index of a cell in a matrix
or returns -1 if rcnames is TRUE and vector y is not within valid range
}
\description{
map the screen coordinates to a cell of a matrix
}
\examples{
if(interactive()){
load("data/x_y.RData")
load("data/x_rcnames.RData")
load("data/x_cal2.RData")
decode(x_y,x_cal2,x_rcnames)
}

}
