% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clickableImageMap.R
\name{defineBounds}
\alias{defineBounds}
\title{defineBounds}
\usage{
defineBounds()
}
\value{
returns a list of numeric xmin xmax ymin ymax defining screen target for exit
}
\description{
use mouse clicks to define bounding box
}
\details{
use in conjunction with exitClick()
}
\examples{
if(interactive()){
defineBounds()
}

}
