% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clickableImageMap.R
\name{doubleClick}
\alias{doubleClick}
\title{doubleClick}
\usage{
doubleClick(tol = 0.001)
}
\arguments{
\item{tol}{numeric tolerance for detecting same position}
}
\value{
returns TRUE if a double click was detected
}
\description{
detect a (left) double click (without moving cursor position)
}
\details{
I realized this is not very useful, as processing is stopped until 2 clicks are detected
}
\examples{
if(interactive()){
doubleClick()
}

}
