% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clickableImageMap.R
\name{exitClick}
\alias{exitClick}
\title{exitClick}
\usage{
exitClick(bounds, y)
}
\arguments{
\item{bounds}{list of numeric xmin xmax ymin ymax defining screen target for exit}

\item{y}{numeric vector of x and y cursor position}
}
\value{
Boolean TRUE if y is within bounds
}
\description{
test position of mouse click to see if user wants to exit
}
\details{
use in conjunction with defineBounds()
}
\examples{
if(interactive()){
load("data/x_bounds.RData")
load("data/x_y.RData")
exitClick(x_bounds,x_y)
}

}
