% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clickableImageMap.R
\name{gtable_replace_grob}
\alias{gtable_replace_grob}
\title{gtable_replace_grob}
\usage{
gtable_replace_grob(gtab, row, new_grob, name)
}
\arguments{
\item{gtab}{a gtable object}

\item{row}{integer target row number within the gtable}

\item{new_grob}{update grob to insert into gtable}

\item{name}{character string entry in the "name" field of gtable$layout}
}
\value{
returns the updated gtable object
}
\description{
replace an existing grob (in a row of a gtable)
with an updated version
}
\examples{
if(interactive()){
load("data/x_gtab.RData")
load("data/x_tab.RData")
load("data/x_rows.RData")
ptabRow<-which(names(x_rows)=="ptabRow")
gtab<-gtable_replace_grob(x_gtab,ptabRow,x_tab,name="ptab")
}
}
