% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clickableImageMap.R
\name{highlight}
\alias{highlight}
\title{highlight}
\usage{
highlight(gtab, color, fontsize)
}
\arguments{
\item{gtab}{a gtable object}

\item{color}{character string representing a color}

\item{fontsize}{integer font size}
}
\value{
returns gtab
}
\description{
invoke highlight() to set highlight font color and size
}
\examples{
if(interactive()){
load("data/x_gtab.RData")
highlight(x_gtab,"red",16)
}

}
