% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clickableImageMap.R
\name{highlightOneCell}
\alias{highlightOneCell}
\title{highlightOneCell}
\usage{
highlightOneCell(gtab, row, col, currentPick)
}
\arguments{
\item{gtab}{a gtable object}

\item{row}{integer row number of cell to highlight}

\item{col}{integer col number of cell to highlight}

\item{currentPick}{Boolean TRUE if this is the most recently chosen cell
and we are to apply special highlighting}
}
\value{
returns gtab, a gtable object
}
\description{
highlight one cell of grob matrix in gtab
}
\examples{
if(interactive()){
load("data/x_mtab.RData")
load("data/x_clickCoord.RData")
highlightOneCell(x_mtab,x_clickCoord[1,"x"],x_clickCoord[1,"y"],FALSE)
}

}
