% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clickableImageMap.R
\name{pullDown}
\alias{pullDown}
\title{pullDown}
\usage{
pullDown(gtab, row, focus)
}
\arguments{
\item{gtab}{a gtable object}

\item{row}{integer target row number within the gtable}

\item{focus}{Boolean if TRUE add emphasis to matrix cell}
}
\value{
returns a list whose components are the generated matrix
and the gtable object
}
\description{
generate and insert a matrix, acting as a pull down menu,
into a gtable object
}
\examples{
if(interactive()){
load("data/x_gtab.RData")
load("data/x_rows.RData")
pullDownRow<-which(names(x_rows)=="pullDownRow")
message<-"select a new value from the pull down menu: "
pd<-pullDown(x_gtab,pullDownRow,grepl("pull down",message))
}

}
