% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clickableImageMap.R
\name{tabify}
\alias{tabify}
\title{tabify}
\usage{
tabify(m, focus = FALSE, clickCoord = NULL)
}
\arguments{
\item{m}{a matrix}

\item{focus}{Boolean if TRUE add emphasis to matrix cell}

\item{clickCoord}{param for highlightOneCell()}
}
\value{
returns the grob containing the matrix
}
\description{
adjust the width and height of a matrix to exactly fill the
grob
}
\examples{
if(interactive()){
load("data/x_m.RData")
t<-tabify(x_m,FALSE,NULL)
}

}
