% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clickableImageMap.R
\name{unhighlight}
\alias{unhighlight}
\title{unhighlight}
\usage{
unhighlight(gtab)
}
\arguments{
\item{gtab}{a gtable object}
}
\value{
returns the return value of highlight()
}
\description{
invoke highlight() to set font color and size to default
}
\examples{
if(interactive()){
load("data/x_gtab.RData")
unhighlight(x_gtab)
}

}
