% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clickableImageMap.R
\name{zlocator}
\alias{zlocator}
\title{zlocator}
\usage{
zlocator(cal, rcnames, bounds)
}
\arguments{
\item{cal}{return value of calibrate()}

\item{rcnames}{parameter passed to decode()}

\item{bounds}{parameter passed to exitClick()}
}
\value{
returns the return value of decode()
}
\description{
wrapper to perform and decode grid.locator()
}
\details{
keeps looping until a valid click is detected
}
\examples{
if(interactive()){
load("data/x_cal.m.RData")
load("data/x_rcnames.RData")
load("data/x_bounds.RData")
zlocator(x_cal.m,x_rcnames,x_bounds)
}

}
